[skip] For polymorphic has_one relationships, optionally passed type to the relationship block (2m)
When plucking for JSON:API, one needs not only the records and data from them, but the information about type you're presenting, as well. There's not really a good way to `SELECT` that out of the associated table, but it's information that the `Relationship::PolymorphicOne` already has available to it from the parent table, and it can pass that type information in to the relationship block to aid in presentation.